requirements.txt

pillow~=10.3.0
torchmetrics
ftfy~=6.1.3
regex~=2024.5.15
setuptools~=68.2.2
matplotlib~=3.8.4
omegaconf~=2.3.0
timm~=0.9.16
requests~=2.31.0
scikit-learn~=1.5.0
opencv-python~=4.9.0.80
scipy~=1.13.0
transformers
datasets~=2.19.1
PyYAML~=6.0.1
huggingface_hub~=0.23.0
hydra-core~=1.3.2
lightning~=2.3.3 
pytorch-lightning~=2.3.3
ipykernel

wandb==0.13.4
einops==0.5.0
pycocotools
scikit-image==0.19.2
joblib==1.2.0
tqdm==4.64.1
notebook==6.5.1
pandas==1.5.1
seaborn==0.12.1
numpy>=1.22
ipdb==0.13.9

git+https://github.com/facebookresearch/segment-anything.git
opencv-python 
onnxruntime 
onnx

mmcv-full


